REGRESSION (r130640 - r130644): TestCookieManager API test is failing https://bugs.webkit.org/show_bug.cgi?id=98738
Reviewed by Philippe Normand.

Source/WebKit2:

For some reason the order of the domains in the array has changed,
but it doesn't really matter, what we want to check is that both
cookies have been accepted and the array contains both domains.

* UIProcess/API/gtk/tests/TestCookieManager.cpp:
(testCookieManagerAcceptPolicy): Check that first and third party
domains are present in the domains array.

Tools:

* Scripts/run-gtk-tests:
(TestRunner): Unskip /webkit2/WebKitCookieManager/accept-policy.

Identifier: 117812@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131876 268f45cc-cd09-0410-ab3c-d52691b4dbfc