Lexer should create 8 bit Identifiers for RegularExpressions and ASCII identifiers https://bugs.webkit.org/show_bug.cgi?id=99855
Reviewed by Filip Pizlo.

Added makeIdentifier helpers that will always make an 8 bit Identifier or make an
Identifier that is the same size as the template parameter.  Used the first in the fast
path when looking for a JS identifier and the second when scanning regular expressions.

* parser/Lexer.cpp:
(JSC::::scanRegExp):
* parser/Lexer.h:
(Lexer):
(JSC::::makeIdentifierSameType):
(JSC::::makeLCharIdentifier):
(JSC::::lexExpectIdentifier):


Identifier: 117892@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131956 268f45cc-cd09-0410-ab3c-d52691b4dbfc