In some float situations, the original layout is wrong and only corrects itself on repaint https://bugs.webkit.org/show_bug.cgi?id=18939
Reviewed by Levi Weintraub.

Source/WebCore:

A <br> that follows collapsing spaces and has clearance set always needs to get a run and a linebox of its own.
This allows the line below it to check for clearance from the <br> if it gets dirtied but the line with the <br>
does not.

Test: fast/block/br-with-clearance-after-collapsing-space.html

* rendering/RenderBlockLineLayout.cpp:
(WebCore::RenderBlock::LineBreaker::nextLineBreak):

LayoutTests:

* fast/block/br-with-clearance-after-collapsing-space-expected.html: Added.
* fast/block/br-with-clearance-after-collapsing-space.html: Added.

Identifier: 117934@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131999 268f45cc-cd09-0410-ab3c-d52691b4dbfc