Viewport errors should be slightly friendlier with regard to ';'. https://bugs.webkit.org/show_bug.cgi?id=100003
Reviewed by Adam Barth.

Source/WebCore:

This patch scans viewport values that cause errors for ';'. If found,
a quick message is appended to the error, noting that semicolons are
not valid separators in viewport contents, and that commas would be
the proper substitute.

No functional changes, just a better error message.

* dom/ViewportArguments.cpp:
(WebCore::reportViewportWarning):

LayoutTests:

* fast/viewport/viewport-128-expected.txt:
* fast/viewport/viewport-133-expected.txt:
* fast/viewport/viewport-134-expected.txt:
* fast/viewport/viewport-54-expected.txt:
* fast/viewport/viewport-65-expected.txt:
* fast/viewport/viewport-67-expected.txt:
* fast/viewport/viewport-77-expected.txt:
* fast/viewport/viewport-warnings-5-expected.txt:
    Updating tests with new expectations.


Identifier: 118106@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132186 268f45cc-cd09-0410-ab3c-d52691b4dbfc