POTFILES.in/.skip need updates for translators https://bugs.webkit.org/show_bug.cgi?id=67580
Reviewed by Xan Lopez.

.:

Now process POTFILES.in during configuration.

* configure.ac:

Source/WebKit/gtk/po:

Move POTFILES to POTFILES.in to satisfy intltools-update and also update
the list of files with translations. Since the way the path is calculated
is different now, we also need to change the gettext invocation in
GNUmakefile.am.

* GNUmakefile.am:
* POTFILES: Removed.
* POTFILES.in: Added.

Identifier: 118204@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132289 268f45cc-cd09-0410-ab3c-d52691b4dbfc