[EFL][WK2][AC] Black screen in web inspector window with AC on https://bugs.webkit.org/show_bug.cgi?id=100168
Patch by Yael Aharon <yael.aharon@intel.com> on 2012-10-24
Reviewed by Kenneth Rohde Christiansen.

When ACCELERATED_COMPOSITING flag is turned on, and using X11,
pass "opengl_x11" engine name to ecore_evas_new when creating a
web inspector window. It is guaranteed to be available, because
we wouldn't be able to create a main window if it wasn't.

(WebKit::WebInspectorProxy::platformCreateInspectorPage):

Identifier: 118244@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132329 268f45cc-cd09-0410-ab3c-d52691b4dbfc