[MathML] Timeouts on linux after r132264 https://bugs.webkit.org/show_bug.cgi?id=100202
Reviewed by Eric Seidel.

Source/WebCore:

When building a stretched operator, such as a large parenthesis or bracket, we need to
check that the extension glyph's height is > 0, to avoid an infinite loop. The 0 height
can occur if the glyph is missing on the host system.

Tested by existing tests.

* rendering/mathml/RenderMathMLOperator.cpp:
(WebCore::RenderMathMLOperator::updateFromElement):

LayoutTests:

The failing tests no longer timeout.

* platform/chromium/TestExpectations:


Identifier: 118280@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132365 268f45cc-cd09-0410-ab3c-d52691b4dbfc