[chromium] Add textureWidth() and textureHeight() to WebVideoFrame https://bugs.webkit.org/show_bug.cgi?id=100042
Patch by John Sheu <sheu@google.com> on 2012-10-24
Reviewed by James Robinson.

It might be the case that the visible portion of a video frame isn't the
entire texture size, e.g. for cropping.  We'll need accessors in
WebVideoFrame to be able to query both the underlying texture dimensions
and the subrect which is supposed to be visible.

This is presently a no-op change.  Once Chromium changes are in, I will
remove the superseded parts of the WebVideoFrame interface.

* chromium/public/WebVideoFrame.h:
(WebKit::WebVideoFrame::visibleRect);
(WebKit::WebVideoFrame::textureSize);

Identifier: 118362@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132448 268f45cc-cd09-0410-ab3c-d52691b4dbfc